pyproject.toml
[project]
name = "devflow-agent"
version = "0.2.0"
description = "Unattended runtime adapter that wires devflow-lite hooks into claude-agent-sdk."
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
  "claude-agent-sdk>=0.1.0",
  "anyio>=4.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/devflow_agent"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]
# ASYNC240 flags pathlib.Path.resolve() which is a non-blocking os.path call,
# not async file I/O — the rule is a false-positive for this usage.
ignore = ["ASYNC240"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
